Extract interfaces for EhcacheClientEntity
