+++ clustered/integration-test/src/test/java/org/ehcache/clustered/CacheManagerLifecycleEhcacheIntegrationTest.java
@@ -32,8 +32,9 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
 import org.ehcache.PersistentCacheManager;
-import org.ehcache.clustered.client.internal.EhcacheClientEntity;
 import org.ehcache.clustered.client.config.builders.ClusteringServiceConfigurationBuilder;
+import org.ehcache.clustered.client.internal.InternalEhcacheClientEntity;
+import org.ehcache.clustered.client.internal.SimpleEhcacheClientEntity;
 import org.ehcache.clustered.common.EhcacheEntityVersion;
 import org.ehcache.config.builders.CacheManagerBuilder;
 import org.ehcache.StateTransitionException;
@@ -79,14 +80,14 @@ public class CacheManagerLifecycleEhcacheIntegrationTest {
 
   @Test
   public void testAutoCreatedCacheManager() throws Exception {
-    assertEntityNotExists(EhcacheClientEntity.class, "testAutoCreatedCacheManager");
+    assertEntityNotExists(InternalEhcacheClientEntity.class, "testAutoCreatedCacheManager");
     PersistentCacheManager manager = newCacheManagerBuilder()
             .with(ClusteringServiceConfigurationBuilder.cluster(CLUSTER.getConnectionURI().resolve("/testAutoCreatedCacheManager")).autoCreate().build())
             .build();
-    assertEntityNotExists(EhcacheClientEntity.class, "testAutoCreatedCacheManager");
+    assertEntityNotExists(InternalEhcacheClientEntity.class, "testAutoCreatedCacheManager");
     manager.init();
     try {
-      assertEntityExists(EhcacheClientEntity.class, "testAutoCreatedCacheManager");
+      assertEntityExists(InternalEhcacheClientEntity.class, "testAutoCreatedCacheManager");
     } finally {
       manager.close();
     }
@@ -98,10 +99,10 @@ public class CacheManagerLifecycleEhcacheIntegrationTest {
     URL xml = CacheManagerLifecycleEhcacheIntegrationTest.class.getResource("/configs/clustered.xml");
     URL substitutedXml = substitute(xml, "cluster-uri", CLUSTER.getConnectionURI().toString());
     PersistentCacheManager manager = (PersistentCacheManager) newCacheManager(new XmlConfiguration(substitutedXml));
-    assertEntityNotExists(EhcacheClientEntity.class, "testAutoCreatedCacheManagerUsingXml");
+    assertEntityNotExists(InternalEhcacheClientEntity.class, "testAutoCreatedCacheManagerUsingXml");
     manager.init();
     try {
-      assertEntityExists(EhcacheClientEntity.class, "testAutoCreatedCacheManagerUsingXml");
+      assertEntityExists(InternalEhcacheClientEntity.class, "testAutoCreatedCacheManagerUsingXml");
     } finally {
       manager.close();
     }
@@ -109,7 +110,7 @@ public class CacheManagerLifecycleEhcacheIntegrationTest {
 
   @Test
   public void testMultipleClientsAutoCreatingCacheManager() throws Exception {
-    assertEntityNotExists(EhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
+    assertEntityNotExists(InternalEhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
 
     final CacheManagerBuilder<PersistentCacheManager> managerBuilder = newCacheManagerBuilder()
             .with(ClusteringServiceConfigurationBuilder.cluster(CLUSTER.getConnectionURI().resolve("/testMultipleClientsAutoCreatingCacheManager")).autoCreate().build());
@@ -123,7 +124,7 @@ public class CacheManagerLifecycleEhcacheIntegrationTest {
       }
     };
 
-    assertEntityNotExists(EhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
+    assertEntityNotExists(InternalEhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
 
     ExecutorService executor = Executors.newCachedThreadPool();
     try {
@@ -134,7 +135,7 @@ public class CacheManagerLifecycleEhcacheIntegrationTest {
       for (Future<PersistentCacheManager> result : results) {
         result.get().close();
       }
-      assertEntityExists(EhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
+      assertEntityExists(InternalEhcacheClientEntity.class, "testMultipleClientsAutoCreatingCacheManager");
     } finally {
       executor.shutdown();
     }
